+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add "Business support" as a team option for users.
+- Add a CSV export for Academies due to transfer in a time period (this work is
+  not publicly viewable)
 
 ### Changed